Changes all intefer types to stdint.h's types

+++ uni-project/config.h
@@ -8,22 +8,23 @@
 #ifndef CONFIG_H_
 #define CONFIG_H_
 
+#include <stdint.h>
 #include "morse_symbols.h"
 
 
-const int kLedPin       = 3;
-const int kRedLedPin_1  = 5;
-const int kRedLedPin_2  = 7;
-const int kSensorPin    = 0;
+const uint8_t kLedPin       = 3;
+const uint8_t kRedLedPin_1  = 5;
+const uint8_t kRedLedPin_2  = 7;
+const uint8_t kSensorPin    = 0;
 
-const int kLightSensorThreshold = 60;   // percent. Higher = more sensitive.
-const int kWaitForPressInterval = 50;   // milliseconds
-const int kDotDashBorder        = 225;  // milliseconds of press
-const int kSerialDataRate       = 9600; // baud
-const int kLoopDelay            = 40;   // milliseconds
-const int kNewWordThreshhold    = 2000; // milliseconds
+const uint8_t kLightSensorThreshold = 60;   // percent. Higher = more sensitive.
+const uint16_t kWaitForPressInterval = 50;   // milliseconds
+const uint16_t kDotDashBorder        = 225;  // milliseconds of press
+const uint32_t kSerialDataRate       = 9600; // baud
+const uint16_t kLoopDelay            = 40;   // milliseconds
+const uint16_t kNewWordThreshhold    = 2000; // milliseconds
 
-const int kEndOfSentenceTimer   = 4000; // milliseconds
+const uint16_t kEndOfSentenceTimer   = 4000; // milliseconds
 
 
 #endif  // CONFIG_H_
