+++ uni-project/led_utils.h
@@ -12,34 +12,34 @@
 /**
  *  Declarations
  */
-void LED_TurnOn(const int pin_num);
-void LED_TurnOff(const int pin_num);
+void LED_TurnOn(const uint8_t pin_num);
+void LED_TurnOff(const uint8_t pin_num);
 
-void LED_Flash(const int pin_num, const unsigned long duration);
-void LED_Blink(const int pin_num, const unsigned long flash_dur, const unsigned long blink_dur);
-void LED_DoubleBlink_Split(const int pin1_num, const int pin2_num, const unsigned long flash_dur, const unsigned long blink_dur);
-void LED_DoubleBlink_Parallel(const int pin1_num, const int pin2_num, const unsigned long flash_dur, const unsigned long blink_dur);
+void LED_Flash(const uint8_t pin_num, const uint32_t duration);
+void LED_Blink(const uint8_t pin_num, const uint32_t flash_dur, const uint32_t blink_dur);
+void LED_DoubleBlink_Split(const uint8_t pin1_num, const uint8_t pin2_num, const uint32_t flash_dur, const uint32_t blink_dur);
+void LED_DoubleBlink_Parallel(const uint8_t pin1_num, const uint8_t pin2_num, const uint32_t flash_dur, const uint32_t blink_dur);
 
 
 /**
  *  Implementations
  */
-void LED_TurnOn(const int pin_num) {
+void LED_TurnOn(const uint8_t pin_num) {
   digitalWrite(pin_num, HIGH);
 }
 
-void LED_TurnOff(const int pin_num) {
+void LED_TurnOff(const uint8_t pin_num) {
   digitalWrite(pin_num, LOW);
 }
 
-void LED_Flash(const int pin_num, const unsigned long duration) {
+void LED_Flash(const uint8_t pin_num, const uint32_t duration) {
   digitalWrite(pin_num, HIGH);
   delay(duration);
   digitalWrite(pin_num, LOW);
 }
 
-void LED_Blink(const int pin_num, const unsigned long flash_dur, const unsigned long blink_dur) {
-  for (unsigned long time = 0; time < blink_dur; time += flash_dur * 2) {
+void LED_Blink(const uint8_t pin_num, const uint32_t flash_dur, const uint32_t blink_dur) {
+  for (uint32_t time = 0; time < blink_dur; time += flash_dur * 2) {
     LED_Flash(pin_num, flash_dur);
     delay(flash_dur);
   }
@@ -47,8 +47,8 @@ void LED_Blink(const int pin_num, const unsigned long flash_dur, const unsigned 
   LED_TurnOff(pin_num);
 }
 
-void LED_DoubleBlink_Split(const int pin1_num, const int pin2_num, const unsigned long flash_dur, const unsigned long blink_dur) {
-  for (unsigned long time = 0; time < blink_dur; time += flash_dur * 2) {
+void LED_DoubleBlink_Split(const uint8_t pin1_num, const uint8_t pin2_num, const uint32_t flash_dur, const uint32_t blink_dur) {
+  for (uint32_t time = 0; time < blink_dur; time += flash_dur * 2) {
     LED_TurnOn(pin1_num);
     LED_TurnOff(pin2_num);
     delay(flash_dur);
@@ -62,8 +62,8 @@ void LED_DoubleBlink_Split(const int pin1_num, const int pin2_num, const unsigne
   LED_TurnOff(pin2_num);
 }
 
-void LED_DoubleBlink_Parallel(const int pin1_num, const int pin2_num, const unsigned long flash_dur, const unsigned long blink_dur) {
-    for (unsigned long time = 0; time < blink_dur; time += flash_dur * 2) {
+void LED_DoubleBlink_Parallel(const uint8_t pin1_num, const uint8_t pin2_num, const uint32_t flash_dur, const uint32_t blink_dur) {
+    for (uint32_t time = 0; time < blink_dur; time += flash_dur * 2) {
     LED_TurnOn(pin1_num);
     LED_TurnOn(pin2_num);
     delay(flash_dur);
